Split comment id sequences into application, overtime and sick note sequences
