+++ src/main/java/org/synyx/urlaubsverwaltung/sicknote/comment/SickNoteCommentEntity.java
@@ -3,13 +3,18 @@ package org.synyx.urlaubsverwaltung.sicknote.comment;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Enumerated;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
+import jakarta.persistence.SequenceGenerator;
 import jakarta.validation.constraints.NotNull;
 import org.hibernate.annotations.OnDelete;
 import org.synyx.urlaubsverwaltung.comment.AbstractComment;
 
 import java.time.Clock;
+import java.util.Objects;
 
 import static jakarta.persistence.EnumType.STRING;
+import static jakarta.persistence.GenerationType.SEQUENCE;
 import static org.hibernate.annotations.OnDeleteAction.CASCADE;
 
 /**
@@ -17,6 +22,12 @@ import static org.hibernate.annotations.OnDeleteAction.CASCADE;
  */
 @Entity(name = "sick_note_comment")
 public class SickNoteCommentEntity extends AbstractComment {
+
+    @Id
+    @Column(name = "id", unique = true, nullable = false, updatable = false)
+    @GeneratedValue(strategy = SEQUENCE, generator = "sick_note_comment_generator")
+    @SequenceGenerator(name = "sick_note_comment_generator", sequenceName = "sick_note_comment_id_seq")
+    private Long id;
 
     @NotNull
     @Column(name = "sick_note_id")
@@ -34,6 +45,14 @@ public class SickNoteCommentEntity extends AbstractComment {
         super(clock);
     }
 
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
     public Long getSickNoteId() {
         return sickNoteId;
     }
@@ -48,5 +67,22 @@ public class SickNoteCommentEntity extends AbstractComment {
 
     public void setAction(SickNoteCommentAction action) {
         this.action = action;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) {
+            return true;
+        }
+        if (o == null || getClass() != o.getClass()) {
+            return false;
+        }
+        final SickNoteCommentEntity that = (SickNoteCommentEntity) o;
+        return null != this.getId() && Objects.equals(id, that.id);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(id);
     }
 }
